% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDataMod.R
\name{call_gdal_mod}
\alias{call_gdal_mod}
\title{Call GDAL}
\usage{
call_gdal_mod(input, method, cellsize)
}
\arguments{
\item{input}{input grid (raster)}

\item{method}{resampling method}

\item{cellsize}{output cell diminsion}
}
\value{
RasterLayer
}
\description{
Call GDAL
}
\keyword{internal}
