% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{checkNetrc}
\alias{checkNetrc}
\title{Check netrc file}
\usage{
checkNetrc(netrcFile = getNetrcPath())
}
\arguments{
\item{netrcFile}{A character. File path to netrc file to check.}

\item{onStaging}{A logical. Whether to check the staging server instead
of the production server.}
}
\value{
logical
}
\description{
Check that there is a netrc file with a valid
entry for urs.earthdata.nasa.gov.
}
\seealso{
\code{\link{writeNetrc}}
}
