% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim_names}
\alias{dim_names}
\title{NetCDF Dimension Names from a path}
\usage{
dim_names(path)
}
\arguments{
\item{path}{a path to a NetCDF file}
}
\value{
a vector of dimension names
}
\description{
Given a path to a NetCDF file, list the diminsions
}
