% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_conus_nwm}
\alias{download_conus_nwm}
\title{Download National Water Model Domain Files}
\usage{
download_conus_nwm(outDir = NULL)
}
\arguments{
\item{outDir}{where to write the files. Again ~17GB worth!}
}
\value{
path to domain file downloads
}
\description{
Download the domain files (CONUS) used in the most
current operational NWM. It will download
~17GB of data so plan accordingly!
}
\details{
This function will download the following national (CONUS) domain
files for the latest version of the National Water Model on NCEP.
\itemize{
\item{"Fulldom_hires_netcdf_1km.nc"}
\item{"geo_em.d01_1km.nc"}
\item{"wrfinput_d01_1km.nc"}
\item{"RouteLink_NHDPLUS.nc"}
\item{"spatialweights_1km_all_basins.nc"}
\item{"GWBUCKPARM_CONUS.nc"}
\item{"soil_veg_properties_LongRange.nc"}
\item{"HYDRO_TBL_2D.nc"}
\item{"WRF_Hydro_NWM_geospatial_data_template_land_GIS.nc"}
}

This function will NOT download:
\itemize{
\item{"Fulldom_hires_netcdf_250m.nc"}("~13GB)
\item{"GWBUCKPARM_CONUS_LongRange.nc"}("92B")
\item{"HYDRO_TBL_2D_LongRange.nc"}{"338MB"}
\item{"nudgingParams.nc"}{"2.4MB"}
\item{"soil_veg_properties_ASM.nc"}{"3GB}
\item{"soil_veg_properties_LongRange.nc"}{"3GB"}
\item{"spatialweights_250m_all_basins.nc"}{"4.5GB"}
}
}
