% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo_grid_proj}
\alias{geo_grid_proj}
\title{Get GeoGrid Projection}
\usage{
geo_grid_proj(path)
}
\arguments{
\item{path}{a path to the local geogrid.nc file}
}
\value{
a PROJ4STRING
}
\description{
From a geogrid file, extract the proj4string.
}
\details{
Extracts projection data from global attributes
}
