% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sequence.R
\name{mask_geogrid_byWS}
\alias{mask_geogrid_byWS}
\title{mask_geogrid_byBasin}
\usage{
mask_geogrid_byWS(RasterLayer, basin)
}
\arguments{
\item{RasterLayer}{GEOGRID as RasterLayer stack}

\item{basin}{Watershed boundary as sf}
}
\value{
a dataframe
}
\description{
Mask GEOGRID by Watershed Boundary. This can be used to perturb LULC within WS boundary or to calculate LULC statistics within WS.
}
