% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nc_to_df}
\alias{nc_to_df}
\title{Read a NetCDF file into a data.frame}
\usage{
nc_to_df(path)
}
\arguments{
\item{path}{a path to a NetCDF}
}
\value{
a data.frame
}
\description{
Read a NetCDF file into a data.frame
}
