% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDataMod.R
\name{output_grid}
\alias{output_grid}
\title{Build output grid}
\usage{
output_grid(input, cellsize)
}
\arguments{
\item{input}{Polygon or WRF-Hydro geogrid}

\item{cellsize}{resolution}
}
\description{
Builds output grid
}
