% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_wt_file}
\alias{read_wt_file}
\title{Re-write of rwfhydro ReadWtFile}
\usage{
read_wt_file(wtFile)
}
\arguments{
\item{path}{path to the weight file}
}
\value{
list of 2 data.frames
}
\description{
reads in a spatial weight file and generates
a list of two dataframes: 1 for the gridded info (data) and 2 for
the poly info (poly)
}
