% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sequence.R
\name{subset_sequence}
\alias{subset_sequence}
\title{Subset sequence wrapper}
\usage{
subset_sequence(
  comid,
  siteID,
  FULLDOMAINDIR,
  outDir,
  nlcdDir = NA,
  methodList = NA
)
}
\arguments{
\item{comid}{comid of outlet reach}

\item{siteID}{siteID for NWIS}

\item{FULLDOMAINDIR}{the value representing "NO DATA", default is NA}

\item{outDir}{output directory (e.g. /home/subsetDOMAINS/)}

\item{nlcdDir}{directory for nlcd dataset}

\item{methodList}{resampling algorithm to use for resampling}
}
\description{
Initiates WrfhydroSubsetter workflow
}
