% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_weights}
\alias{subset_weights}
\title{Re-write of rwrfhydro::SubsetWts}
\usage{
subset_weights(wts, rlids, istart, iend, jstart, jend)
}
\arguments{
\item{wts}{Weight file list object}

\item{rlids}{Vector of link IDs to keep in the subsetted output}

\item{istart}{Start index for the grid in the i (x) direction}

\item{iend}{End index for the grid in the i (x) direction}

\item{jstart}{Start index for the grid in the j (y) direction}

\item{jend}{End index for the grid in the j (y) direction}
}
\value{
List of 2 dataframes
}
\description{
SubsetWts takes a weight file object (as read in from ReadWtFile)
and subsets the data and poly dataframes based on a user-defined set of link
IDs and user-defined start and end indices. Returns a list of 2 dataframes
for the subsetted data and poly objects.
}
