% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleDataMod.R
\name{summarize.cells}
\alias{summarize.cells}
\title{Summarize Output Grid Cells}
\usage{
\method{summarize}{cells}(input, output, no_data = NA)
}
\arguments{
\item{input}{an input raster to resample}

\item{output}{the output grid to resample to. \code{sf} object}

\item{no_data}{the value representing "NO DATA", default is NA}
}
\value{
a tibble
}
\description{
Summarizies the percentage of each output grid cell that is covered by each disiticnt class in the input grid
}
