% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldasR.R
\docType{data}
\name{timestep}
\alias{timestep}
\title{Download CONUS NLDAS Forcing Files}
\format{
An object of class \code{character} of length 2.
}
\usage{
timestep
}
\arguments{
\item{netrcFile}{}
}
\description{
Download CONUS wide NLDAS forcing files for local use. Hourly files are downloaded for each day between the requested start and end date. One year of data requires ~13 GB of disk space. Each file is ~1.8 MB.
}
\details{
This function will download the following national or subset NLDAS forcing files for the latest version of the National Water Model on NCEP.
\itemize{
\item{"Fulldom_hires_netcdf_1km.nc"}
\item{"geo_em.d01_1km.nc"}
\item{"wrfinput_d01_1km.nc"}
\item{"RouteLink_NHDPLUS.nc"}
\item{"spatialweights_1km_all_basins.nc"}
\item{"GWBUCKPARM_CONUS.nc"}
\item{"soil_veg_properties_LongRange.nc"}
\item{"HYDRO_TBL_2D.nc"}
\item{"WRF_Hydro_NWM_geospatial_data_template_land_GIS.nc"}
}
}
\keyword{datasets}
