% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_nc}
\alias{update_nc}
\title{Update the attribute of NetCDF file}
\usage{
update_nc(path, df)
}
\arguments{
\item{path}{Directory of NetCDF file to be updated}

\item{df}{R dataframe  of link IDs to keep in the subsetted output}
}
\value{
Nothing
}
\description{
Update the NetCDF file in path with target dataframe.
}
