% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_names}
\alias{var_names}
\title{NetCDF Variable Names from a path}
\usage{
var_names(path)
}
\arguments{
\item{path}{a path to a NetCDF file}
}
\value{
a vectore of variable names
}
\description{
Given a path to a NetCDF file, list the variables
}
