% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc_utils.R
\name{writeNetrc}
\alias{writeNetrc}
\title{Write netrc file}
\usage{
writeNetrc(
  login,
  password,
  machine = "urs.earthdata.nasa.gov",
  netrcFile = getNetrcPath(),
  overwrite = FALSE
)
}
\arguments{
\item{login}{A character. Email address used for logging in on DataSpace.}

\item{password}{A character. Password associated with the login.}

\item{netrcFile}{A character. A path to where the netrc file should be written.
By default will go to your home directory, which is advised}

\item{overwrite}{A logical. overwrite the existing netrc file?}
}
\value{
A character vector containing the netrc file path
}
\description{
Write a netrc file that is valid for accessing urs.earthdata.nasa.gov
}
\details{
The database is accessed with the user's credentials.
A netrc file storing login and password information is required.
See \href{https://urs.earthdata.nasa.gov/}{here}
for instruction on how to register and set DataSpace credential.
}
\examples{
# First, create an account in the DataSpace App and read the terms of use
# Next, create a netrc file using writeNetrc()
writeNetrc(
  login = "XXX@email.com",
  password = "yourSecretPassword"
)
# Specify `netrcFile = getNetrcPath()` to write netrc in the default path
}
\seealso{
\code{\link{connectDS}} \code{\link{checkNetrc}}
}
